requirements.txt
django==2.1.2
django-cors-headers==2.4.0
djangorestframework==3.9.0
django-annoying==0.10.4
django-enumchoicefield==1.0.2
django-apscheduler==0.2.13